requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
sqlparse
python-jose
passlib[bcrypt]
python-multipart
pydantic-settings
